requirements.txt
flask==3.1.0
flask-cors==5.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
apscheduler==3.10.4

# Optional: choose your email provider
# sendgrid==6.11.0      # Uncomment for SendGrid
# boto3==1.35.0          # Uncomment for AWS SES
